Implement each function
The interesting thing about this implementation is the ease in which a
single method can handle two different argument types. A simple one line
conditional that checks if it is an array handles that case, otherwise it
is treated as an object. In other languages like Java we may need to
implement two different methods each with its own method header
specifying different accepted arguments.