Implement uniq and map
My initial idea to implement uniq was to have a javascript object store
the elements that have already been seen as keys. That way, checking to
see if we have already seen an element while traversing our array would
be done in constant time. I had trouble accomplishing this though
because of the way keys are stringified. So I settled for a nested for
loop.

Map was extreamly easy to implement. I could use the _.each function I
declared earlier instead of a for loop, but I feel more comfortable
with the for loop right now. I'm not sure if that is good or bad.