Implement reduce function
In this case I believe it to be more effective to use a standard for
loop rather than a forEach loop because it gives me more flexability in
determining the starting value.

I use for loop logic (create value i, continue while it is within range,
increment i) but do it in a while loop for readability and
maintainability. I increment the count using

        "accumulator = iterator(accumulator, collection[i++])"

the expression "collection[i++]" means to index into this and retrieve the value,
THEN increment i. This behavior is possible because both indexing and
incrementing are technically considered operators, and indexing has a
higher precedence than incrementing.