Implement some function
Lessons Learned:
When you try to compare a non-boolean type to a boolean it always
evaluates to false.

"string" == true
>>>false

"stirng" == false
>>>false

undefined == false
>>>false

undefined == true
>>>false

This is in direct opposition to the notion that all values evaluate to
booleans. These values need to be converted to booleans to work
correctly.

Boolean("string") == true
>>>true

Boolean(undefined) == false
>>>true