Implement memoize function
Lessons Learned:
The outter function can be used to retain state between function calls.
The variables held in the outter function can therefore be viewed as
metadata.

We are able to do assignment within a logical assignment so long as we
wrap the inner assignment inside parenthesis; this makes it so operator
precedence does not overlap between assignments and conditionals.

Using or operator (||) will return first truthy value

Using and operator (&&) will return last truthy value

After EVERY property assignment, the property value is returned. This is
useful in cases like this where we need to store something and then use
that value right after.

Use 'git commit --amend --no-edit' when you need to add to the last
commit but do not need to edit message