Implement delay function
Lessons Learned:

Array.prototype.slice.call(arguments) can not only be used to extract
the arguments list into an actual array, but we can even use it to
extract the number of arguments we want:
  var arr = Array.prototype.slice.call  (arguments, 2);

This can be useful in the future when we are dealing with wrapper
functions in the future and may need to differentiate between the
arguments that are meant to be used to initialize the wrapper function
itself, and the arguments which are meant to be passed into the wrapper
function as arguments.

A useful way to think of call is to think of the Ruby method. The two
are slightly different because in Ruby we do not set the context.
Nonetheless the method exist in both.

function.call(args) is a common pattern

The thing that is exciting about all these functions is that they are
all so universal and potent.

I think the difference between:
var func = function (){}

and:
function func(){}

is that using the function expression is meant for functions which code
bodies are going to be executed, as where the normal function
declaration syntax is meant for constructors to create objects. I do not
know this for sure, but it seems to be a reoccuring theme I am noticing.