Implement shuffle function
This shuffle function generates a random number between 0 and the size
of the array. If the position of the generated index is already taken,
then the function uses linear probing to find the next avalible index.

We use:
  for (i = 0;arr[i] !== undefined;i++);
for our linear probing. Note the semi colon at the end. This makes it so
we skip the code execution block. We do not need the code block at the
end because there is no point in testing i if the element at the
position is undefined because the loop does that for us. We if did it a
second time it would be wasted processor time.